add tupfile for compiling
This tupfile only builds the test application __test_server.cpp. In a
later stage this will be moved to a tup variant "test" (similar to a
make target) and other variants to build the library as a proper library
will be added.
Also any build output should be moved to a proper build directory.